更新舵机控制器的寄存器定义，重命名位置和速度PID参数，添加新的获取和设置函数，优化EEPROM配置加载逻辑，确保更好的代码可读性和维护性。

+++ lib/servo/src/servo.h
@@ -94,7 +94,7 @@ class Servo {
   float present_load_;
   float present_voltage_;
   float present_current_;
-  float present_temperature_;
+  uint8_t present_temperature_;
 
   uint8_t error_status_;
   bool moving_;
